refactor: move ValueOf and others
