+++ src/object/ValueOf.ts
