+++ src/object/index.ts
@@ -26,4 +26,5 @@ export * from './reduceKey'
 export * from './replaceProperty'
 export * from './Required'
 export * from './someKey'
-
+export * from './typeOverrideIncompatible'
+export * from './ValueOf'
