+++ src/object/typeOverrideIncompatible.spec.ts
