+++ src/object/typeOverrideIncompatible.ts
@@ -1,4 +1,4 @@
-import { ANotB, AnyRecord } from './object'
+import { ANotB, AnyRecord } from '.'
 
 export function typeOverrideIncompatible<A extends AnyRecord>() {
   return function <B extends AnyRecord>(source: B, override: ANotB<A, B>): A {
